hw debug API: aarch64 sw break, single step tests
Ported software breakpoint test to aarch64 and moved single
stepping test from x86-only to arch-independent.
